add v1 wardriving poc w/ splash screen graphics

+++ src/Wardriver/src/Screen.h
@@ -5,5 +5,5 @@ class Screen {
         Screen();
         static void init();
         static void drawSplash(uint8_t sec);
-        static void drawMockup(char* gpscoords, char* time, uint8_t icon1, uint8_t icon2, uint8_t icon3, uint8_t icon4, uint8_t icon5, uint8_t icon6, char* message);
+        static void drawMockup(char* gpscoords, char* time, uint8_t icon1, uint16_t icon2, uint8_t icon3, uint8_t icon4, uint8_t icon5, uint8_t icon6, char* message);
 };
